third_party/WebKit/LayoutTests/fast/dom/shadow/input-color-in-content.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>
<div id="host" tabindex="1">
  <input id="target" type="color" value="#000000">
</div>
<script>

shadow = host.createShadowRoot();
shadow.innerHTML = "<content></content>";
jsTestIsAsync = true;

function waitUntilShown(then)
{
    if (!testRunner.isChooserShown())
        return window.setTimeout(function() { waitUntilShown(then); }, 0);
    then();
}

function ensureOpeningFor(count, then)
{
    shouldBeTrue("testRunner.isChooserShown()");
    if (count)
        return window.setTimeout(function() { ensureOpeningFor(count -1, then); }, 0);
    then();
}

if (window.eventSender) {
   shouldBeFalse("testRunner.isChooserShown()");
   eventSender.dragMode = false;
   eventSender.mouseMoveTo(target.offsetLeft, target.offsetTop + target.offsetHeight / 2);
   eventSender.mouseDown();
   eventSender.mouseUp();

   waitUntilShown(function() {
      testRunner.setWindowIsKey(false); // This triggers a style recalc.
      ensureOpeningFor(10, finishJSTest); // Try several times to ensure nothing wrong happens asynchronously.
   });
}

</script>
</body>
</html>
